pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-task"
version = "0.1.1"
description = "Task-Specific AI Agent Platform for Cursor AI"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["ai", "cursor", "cli", "agent"]
authors = [
    { name = "MicroWise AI", email = "contact@microwiseai.dev" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "cleo>=2.0.0",      # CLI framework
    "appdirs>=1.4.0",   # App directory management
    "rich>=13.0.0",     # Rich terminal output
    "pyyaml>=6.0",      # YAML file handling
    "requests>=2.28.0", # HTTP client
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://taskhub.microwiseai.dev"
Repository = "https://github.com/microwiseai/agent-task"
Documentation = "https://taskhub.microwiseai.dev/docs"

[project.scripts]
agent-task = "agent_task.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["agent_task"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
select = ["E", "F", "B", "I"]
ignore = []
line-length = 88
target-version = "py38" 